requirements.txt
sqlalchemy
psycopg2-binary
fastapi
uvicorn
crewai
langchain
pydantic
python-dotenv
pandas
reportlab
streamlit
requests
langchain-openai
